added persistence for prefault vectors
Summary:
Prefault vectors are currently wiped out on module unload,
but are necessary for prefaulting with a cold buffer cache. Add
export and import routines for the vectors to the SLS.

Test Plan: TODO (already passing, just a formality)

Reviewers: ali, ryan, tavianator

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D688

+++ sls/sls_partition.h
@@ -37,10 +37,10 @@ struct slspart_serial {
 	uint64_t sspart_oid;
 	struct sls_attr sspart_attr;
 	uint64_t sspart_epoch;
-	int sspart_proc;
-	int sspart_td;
+	uint64_t sspart_proc;
+	uint64_t sspart_td;
 	int sspart_flags;
-	int sspart_sockid;
+	uint64_t sspart_sockid;
 };
 
 /**/
